requirement.txt
Django==3.1.5
phonenumbers==8.12.16
Pillow==8.1.0
pipenv==2020.11.15
python-dateutil==2.8.1
python-dotenv==0.15.0
requests==2.25.1
django-import-export==2.5.0
django-phonenumber-field==5.2.0